refactor(storefront): clean up
