+++ apps/storefront/src/services/lazy-loaders/address.ts
@@ -6,7 +6,7 @@ import {
   emptyAddressService,
   isSaleorConfigured,
 } from "../utils/empty-services";
-import { getRequiredSaleorApiUrl } from "./required-env";
+import { getRequiredSaleorApiUrl } from "../utils/required-env";
 
 /**
  * Creates a lazy loader for the address service (Saleor-backed, with an empty
