+++ apps/storefront/src/services/lazy-loaders/collection.ts
@@ -6,7 +6,7 @@ import {
   emptyCollectionService,
   isSaleorConfigured,
 } from "../utils/empty-services";
-import { getRequiredSaleorApiUrl } from "./required-env";
+import { getRequiredSaleorApiUrl } from "../utils/required-env";
 
 /**
  * Creates a lazy loader for the collection service (Saleor-backed, with an empty
