+++ apps/storefront/src/services/lazy-loaders/marketplace.ts
@@ -7,7 +7,7 @@ import {
   emptyMarketplaceService,
   isSaleorConfigured,
 } from "../utils/empty-services";
-import { getRequiredSaleorApiUrl } from "./required-env";
+import { getRequiredSaleorApiUrl } from "../utils/required-env";
 
 /**
  * Creates a lazy loader for the marketplace service (Saleor-backed, with an
